backend/requirements.txt
# Requirements for deploying the backend to Render
# Pinned to the versions currently in the backend venv
fastapi==0.127.0
uvicorn==0.40.0
pymongo==4.15.5
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
pydantic==2.12.5
starlette==0.50.0
anyio==4.12.0
dnspython==2.8.0
motor==3.7.1
websockets
rsa==4.9.1
cryptography==46.0.3
google-api-python-client==2.86.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
python-multipart==0.0.6
